feat: detailed invoices download added

+++ api/src/converter/static/InstallmentStaticConverter.py
@@ -1,4 +1,4 @@
-from python_helper import DateTimeHelper, ObjectHelper, ReflectionHelper
+from python_helper import DateTimeHelper, ObjectHelper
 from python_framework import StaticConverter
 
 from constant import InstallmentConstant
@@ -11,7 +11,7 @@ def overrideDefaultValues(instance, objectKeys=None):
     instance.value = StaticConverter.getValueOrDefault(instance.value if ObjectHelper.isNone(instance.value) else float(instance.value), InstallmentConstant.DEFAULT_VALUE)
     instance.installmentAt = StaticConverter.getValueOrDefault(DateTimeHelper.of(dateTime=instance.installmentAt), DateTimeHelper.now())
     instance.order = StaticConverter.getValueOrDefault(instance.order if ObjectHelper.isNone(instance.order) else int(instance.order), InstallmentConstant.DEFAULT_ORDER)
-    instance.installments = StaticConverter.getValueOrDefault(instance.installments if ObjectHelper.isNone(instance.installments) else float(instance.installments), InstallmentConstant.DEFAULT_INSTALLMENTS)
+    instance.installments = StaticConverter.getValueOrDefault(instance.installments if ObjectHelper.isNone(instance.installments) else int(instance.installments), InstallmentConstant.DEFAULT_INSTALLMENTS)
     instance.status = StaticConverter.getValueOrDefault(InstallmentStatus.map(instance.status), InstallmentConstant.DEFAULT_STATUS)
     return instance
